Insight_file(Before yarn build)/backendFinal/requirements.txt
asgiref==3.3.1
cycler==0.10.0
Django==3.1.5
django-crispy-forms==1.10.0
django-filter==2.4.0
djangoadmin==2.2.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djangotools==2.2.2
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.5
pandas==1.2.1
patsy==0.5.1
Pillow==8.1.0
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
scipy==1.6.0
six==1.15.0
sqlparse==0.4.1
statsmodels==0.12.1
